Extract dta-daemon module from tools-android (#23)
Create a new dta-daemon module for DTA-specific orchestration logic,
separating it from the generic Android tooling in tools-android.

Moved to dta-daemon (with package rename to io.yamsergey.dta.daemon):
- cdp/ — CDP client, watcher, accessibility inspector, network monitor
- sidekick/ — SidekickClient, SidekickConnectionManager, SSE listener
- scroll/ — scroll screenshot capture

tools-android retains general Android tooling: ADB, Gradle resolution,
view hierarchy, screenshots, vector conversion.

Dependency: dta-daemon → tools-android (one-way, no circular).
