+++ dta-daemon/src/main/java/io/yamsergey/dta/daemon/cdp/CdpAccessibilityInspector.java
@@ -1,4 +1,4 @@
-package io.yamsergey.dta.tools.android.cdp;
+package io.yamsergey.dta.daemon.cdp;
 
 import tools.jackson.databind.JsonNode;
 import tools.jackson.databind.ObjectMapper;
