+++ dta-daemon/src/main/java/io/yamsergey/dta/daemon/cdp/CdpWatcherManager.java
@@ -1,6 +1,6 @@
-package io.yamsergey.dta.tools.android.cdp;
+package io.yamsergey.dta.daemon.cdp;
 
-import io.yamsergey.dta.tools.android.inspect.compose.SidekickClient;
+import io.yamsergey.dta.daemon.sidekick.SidekickClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
