+++ dta-daemon/src/main/java/io/yamsergey/dta/daemon/cdp/ChromeDevToolsClient.java
@@ -1,4 +1,4 @@
-package io.yamsergey.dta.tools.android.cdp;
+package io.yamsergey.dta.daemon.cdp;
 
 import tools.jackson.core.JacksonException;
 import tools.jackson.databind.JsonNode;
