+++ dta-daemon/src/main/java/io/yamsergey/dta/daemon/cdp/ChromeSocketDiscovery.java
@@ -1,4 +1,4 @@
-package io.yamsergey.dta.tools.android.cdp;
+package io.yamsergey.dta.daemon.cdp;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
