+++ dta-daemon/src/main/java/io/yamsergey/dta/daemon/scroll/ImageOverlapDetector.java
@@ -1,4 +1,4 @@
-package io.yamsergey.dta.tools.android.inspect.scroll;
+package io.yamsergey.dta.daemon.scroll;
 
 import java.util.Optional;
