+++ settings.gradle
@@ -22,4 +22,5 @@ include 'tools-android-cli'
 include 'dta-cli'
 include 'dta-sidekick'
 include 'dta-mcp'
+include 'dta-daemon'
 include 'dta-server'
